cli: move db, client_db, restoration_db, secretstore_db to a separate mod
